requirements.txt
# Requirements for OCR Data Generator

# Core dependencies
Pillow>=9.0.0          # Image processing
numpy>=1.20.0          # Numerical computations
opencv-python>=4.5.0   # Advanced image operations
tqdm>=4.60.0           # Progress bars

# Optional dependencies for extended functionality
# matplotlib>=3.3.0    # Visualization (optional)
# pandas>=1.3.0        # Data manipulation (optional)
# scikit-image>=0.18.0 # Additional image processing (optional)

# Development dependencies (for contributors)
# pytest>=6.0.0        # Testing
# black>=21.0.0         # Code formatting
# flake8>=3.8.0         # Linting